+++ ref-impl/src/com-api/com-dll/utf8.cpp
@@ -170,7 +170,7 @@ size_t u8stowcs( wchar_t *pw, const char *pu, size_t count )
 	while( *pu && clen<count )
 	{
 		int ulen=u8towc( &pw[clen], pu, 1 );
-		if( ulen<0 ) return -1;
+		if( ulen<0 ) return (size_t) -1;
 		else { clen++; pu+=ulen; }
 	}
 	if( '\0'==*pu && clen<count) pw[clen++]=L'\0';
